refactor(frontend): shared tooltip code
Move shared tooltip code into `lib/map/tooltip`.
