+++ frontend/src/data-layers/droughtOptions/DroughtOptionsHoverDescription.tsx
@@ -1,5 +1,5 @@
 import { Typography } from '@mui/material';
-import { DataItem } from 'details/features/detail-components';
+import { DataItem } from 'lib/map/tooltip/detail-components';
 import { VectorHoverDescription } from 'lib/data-map/types';
 import { FC, useMemo } from 'react';
 import { useRecoilValue } from 'recoil';
@@ -7,7 +7,7 @@ import {
   droughtOptionsColorSpecState,
   droughtOptionsFieldSpecState,
 } from 'data-layers/droughtOptions/state/layer';
-import { DataDescription } from 'app/map/tooltip/DataDescription';
+import { DataDescription } from 'lib/map/tooltip/DataDescription';
 
 export const DroughtOptionsHoverDescription: FC<VectorHoverDescription> = ({
   target,
