+++ frontend/src/data-layers/marine/MarineHoverDescription.tsx
@@ -1,9 +1,9 @@
 import { Typography } from '@mui/material';
-import { DataItem } from 'details/features/detail-components';
+import { DataItem } from 'lib/map/tooltip/detail-components';
 import { VectorHoverDescription } from 'lib/data-map/types';
 import startCase from 'lodash/startCase';
 import { FC } from 'react';
-import { ColorBox } from 'app/map/tooltip/content/ColorBox';
+import { ColorBox } from 'lib/map/tooltip/content/ColorBox';
 
 import { MARINE_HABITATS_LOOKUP } from './domains';
 import { habitatColorMap } from './state/layer';
