+++ frontend/src/data-layers/terrestrial/TerrestrialHoverDescription.tsx
@@ -1,10 +1,10 @@
 import { Typography } from '@mui/material';
-import { DataItem } from 'details/features/detail-components';
+import { DataItem } from 'lib/map/tooltip/detail-components';
 import { VectorHoverDescription } from 'lib/data-map/types';
 import startCase from 'lodash/startCase';
 import { FC } from 'react';
-import { DataDescription } from 'app/map/tooltip/DataDescription';
-import { ColorBox } from 'app/map/tooltip/content/ColorBox';
+import { DataDescription } from 'lib/map/tooltip/DataDescription';
+import { ColorBox } from 'lib/map/tooltip/content/ColorBox';
 
 import { terrestrialSlope, terrestrialElevation } from './color-maps';
 import { landuseColorMap } from './state/layer';
