+++ frontend/src/details/solutions/SolutionsSidebarContent.tsx
@@ -1,10 +1,10 @@
 import { List, ListItem, Typography } from '@mui/material';
 import { terrestrialSlope, terrestrialElevation } from 'data-layers/terrestrial/color-maps';
 import { MARINE_HABITATS_LOOKUP } from 'data-layers/marine/domains';
-import { DataItem } from 'details/features/detail-components';
+import { DataItem } from 'lib/map/tooltip/detail-components';
 import { colorMap } from 'lib/color-map';
 import startCase from 'lodash/startCase';
-import { ColorBox } from 'app/map/tooltip/content/ColorBox';
+import { ColorBox } from 'lib/map/tooltip/content/ColorBox';
 import { FC } from 'react';
 import { habitatColorMap } from 'data-layers/marine/state/layer';
 import { landuseColorMap } from 'data-layers/terrestrial/state/layer';
