+++ frontend/src/lib/map/tooltip/content/ColorBox.tsx
