+++ frontend/src/lib/map/tooltip/detail-components.tsx
